fixed some compiler warnings

+++ src/sequencer/sequences/RelatTimedSequence.cpp
@@ -7,13 +7,14 @@
 RelatTimedSequence::RelatTimedSequence(const std::vector<std::pair<uint64_t, ECSState &>>& seq) :
         RelatTimedSequence(seq, 0) {}
 
-RelatTimedSequence::RelatTimedSequence(const std::vector<std::pair<uint64_t, ECSState &>>& seq, int index) :
+RelatTimedSequence::RelatTimedSequence(const std::vector<std::pair<uint64_t, ECSState &>>& seq, size_t index) :
         waitTime(seq.at(index).first), //use .at() instead of [] for bounds checking
         storedState(seq.at(index).second)
 {
     if (index >= seq.size()-1) {
         this->nextSeq = std::unique_ptr<ISequence>(nullptr);
     } else {
+        //can't use make_unique, constructor is private
         this->nextSeq = std::unique_ptr<ISequence>(new RelatTimedSequence(seq, index + 1));
     }
 }
